refactor(Query) extract combinators to Schema::Mask; break legacy mask wrap
